+++ src/requirements/__test__/requirement-graph.test.ts
@@ -1,4 +1,5 @@
-import RequirementFulfillmentGraph, { CourseWithUniqueId } from '../graph';
+import RequirementFulfillmentGraph from '../graph';
+import { CourseWithUniqueId } from '../graph/types';
 
 const CS3410 = { uniqueId: 3410 };
 const CS3420 = { uniqueId: 3420 };
