+++ src/requirements/graph/processor/add-arbitrary-opt-in-choices.ts
@@ -1,4 +1,4 @@
-import GraphProcessor from '.';
+import GraphProcessor from './definition';
 import RequirementFulfillmentGraph from '..';
 import { CourseForRequirementGraph } from '../types';
