+++ src/requirements/graph/processor/add-basic-user-fulfillment-choices.ts
@@ -1,4 +1,4 @@
-import GraphProcessor from '.';
+import GraphProcessor from './definition';
 import RequirementFulfillmentGraph from '..';
 import { CourseForRequirementGraph } from '../types';
